Reorganize widgets in widgets and smaller components
