+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import './screens/home.dart';
+import 'screens/home.dart';
 
 void main() => runApp(AppMain());
